refactor: Split UniversityProfile into 2 parts, public and not public cache. Implement caching for public part. Change type for universityId param in controllers and services to short according to entity id type. Make that avatar url is now publicly available and not specially generated for each call.

+++ backend/src/main/java/com/go_exchange_easier/backend/core/domain/university/UniversityApi.java
@@ -50,7 +50,7 @@ public interface UniversityApi {
                             schema = @Schema(implementation = ApiErrorResponse.class)))
     })
     ResponseEntity<UniversityProfile> getProfile(
-            @PathVariable Integer universityId,
+            @PathVariable Short universityId,
             @AuthenticationPrincipal AuthenticatedUser authenticatedUser);
 
     @GetMapping("/{universityId}/reviews")
